requirements.txt
jinja2
python-slugify
requests
beautifulsoup4
lxml
feedparser
